requirements.txt
fastapi==0.115.0
uvicorn==0.22.0
psycopg2-binary==2.9.6
sqlalchemy==2.0.17
alembic==1.11.1
python-dotenv==1.0.0
loguru==0.7.0
redis==4.3.3
celery==5.2.7
passlib==1.7.4
pyjwt==2.3.0
pydantic==2.8.2
pydantic-settings==2.0.3
